backend/requirements.txt
flask
flask-cors
python-dotenv
PyMuPDF
python-docx
